changed totalRewardPoints from Integer to Double, changed logic for more precise points and removed some unnecessary test cases

+++ RewardPoints/CustomerRewards/src/main/java/com/charter/rewardpoints/dto/RewardCalculationResponse.java
@@ -1,9 +1,6 @@
 package com.charter.rewardpoints.dto;
-
 import java.util.List;
-
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
@@ -19,7 +16,7 @@ public class RewardCalculationResponse {
     private List<MonthlyRewardPoints> monthlyRewardPoints;
 
     @Schema(description = "Total reward points for the customer")
-    private Integer totalRewardPoints;
+    private Double totalRewardPoints;
 
     @Schema(description = "List of purchase details of a customer")
     private List<CustomerPurchaseDetails> customerPurchaseDetails;
